chore: move sliding piece move gen to new struct
Created a new sliding piece attack struct that will handle only sliding piece attacks. This way we can abstract the backend method used to generate these attack tables and look them up.
bench: 2517656
